Module Datatype_sig.Sum3

Parameters

module BA : S

module BB : S

module BC : S

Signature

type t = (BA.t, BB.t, BC.t) sum3

val equal : t -> t -> bool

Any notion of equality is allowed, as long as it is an equivalence relation, and that if a == b, then equal a b.

val compare : t -> t -> int

compare is a total order, and should be compatible with equal.

val hash : t -> int

hash requires that equal values have the same hash.

val pretty : Format.formatter -> t -> unit